Modified database table name
Renamed the "dashboards" table name to "logs" and renamed both the model
and controller. Added a new "members" table to replace the "users" table
for regular unit members who will be added by the admin for the
application. These "members" will be the personnel that will operate the
UAVs and so will not have direct access to the dashboard unless they are
given the permission by the unit admin(who will grant them commander
status).